refactor: Implement final kernel architecture and dynamic app system
This commit completes a major architectural refactoring, performed as a result of an interactive and iterative process with the user. The initial exploratory request evolved into a collaborative effort to redesign the application's core for better stability and scalability.

- **Kernel Encapsulation:** All core OS shell logic (main component, UI shell, hooks, contexts, types, constants) is now fully encapsulated within a new `window/` directory. The root directory is now clean.

- **Dynamic App System:** A dynamic app discovery system, driven by `.app` JSON files and a backend API, has been implemented and is used by all relevant UI components.

- **Bug Fixes & Verification:** All import paths have been corrected after the refactoring, and all identified bugs have been resolved.
