+++ components/apps/SettingsApp.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
-import { AppDefinition, AppComponentProps } from '../../../window/types';
-import { SettingsIcon } from '../../../window/constants';
+import { AppDefinition, AppComponentProps } from '../../window/types';
+import { SettingsIcon } from '../../window/constants';
 import { useTheme, themes } from '../../window/theme';
 
 const SettingsApp: React.FC<AppComponentProps> = ({ setTitle, onWallpaperChange }) => {
