+++ components/apps/ThemeApp.tsx
@@ -1,7 +1,7 @@
 
 import React, { useEffect } from 'react';
-import { AppDefinition, AppComponentProps } from '../../../window/types';
-import { ThemeIcon } from '../../../window/constants';
+import { AppDefinition, AppComponentProps } from '../../window/types';
+import { ThemeIcon } from '../../window/constants';
 import { themes, useTheme } from '../../window/theme';
 
 const ThemeApp: React.FC<AppComponentProps> = ({ setTitle, onWallpaperChange }) => {
